# docs/commercialization/openapi/open-v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "YINYU CTF Platform Open API",
    "description": "YINYU CTF Platform external API contract",
    "version": "v1"
  },
  "paths": {
...
    "/api/open/v1/teamlab/capabilities": {
      "get": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Get TeamLab capabilities",
        "description": "Returns the topology schema and feature capabilities supported by this platform version.",
        "operationId": "OpenTeamLabTopologies_Capabilities",
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamLabCapabilitiesModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/topologies": {
      "post": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Create a topology",
        "description": "Queues creation of a reusable TeamLab topology draft.",
        "operationId": "OpenTeamLabTopologies_Create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenCreateTeamLabTopologyModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "List topologies",
        "description": "Returns a cursor-paginated list of topologies visible to the current API token owner.",
        "operationId": "OpenTeamLabTopologies_List",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100.0,
              "minimum": 1.0
            },
            "x-position": 1
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTeamLabTopologyPageModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/topologies/{topologyId}": {
      "get": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Get a topology",
        "description": "Returns the complete editable topology definition.",
        "operationId": "OpenTeamLabTopologies_Get",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTeamLabTopologyDetailModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Update a topology",
        "description": "Queues replacement of the editable topology definition.",
        "operationId": "OpenTeamLabTopologies_Update",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenUpdateTeamLabTopologyModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Delete a topology",
        "description": "Queues deletion of a topology that is no longer in use.",
        "operationId": "OpenTeamLabTopologies_Delete",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/topologies/{topologyId}/validate": {
      "post": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Validate a topology",
        "description": "Validates topology structure, addressing, connectivity, assets, and deployment constraints without publishing it.",
        "operationId": "OpenTeamLabTopologies_Validate",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamLabValidationResultModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/topologies/{topologyId}/releases": {
      "post": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "Publish a topology release",
        "description": "Validates and queues creation of an immutable topology release for runtime deployment.",
        "operationId": "OpenTeamLabTopologies_Publish",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishTeamLabTopologyModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "TeamLab - Topologies"
        ],
        "summary": "List topology releases",
        "description": "Returns immutable releases for a topology using cursor pagination.",
        "operationId": "OpenTeamLabTopologies_ListReleases",
        "parameters": [
          {
            "name": "topologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100.0,
              "minimum": 1.0
            },
            "x-position": 2
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTeamLabReleasePageModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ExternalApiProblemDetailsModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "properties": {
              "code": {
                "type": "string"
              },
              "traceId": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "TeamLabAssetKind": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Docker",
          "Vm"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "TeamLabCapabilitiesModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "topologySchemaVersions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "assetKinds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabAssetKind"
            }
          },
          "networkModel": {
            "type": "string"
          },
          "features": {
            "$ref": "#/components/schemas/TeamLabFeatureCapabilitiesModel"
          },
          "limits": {
            "$ref": "#/components/schemas/TeamLabContractLimitsModel"
          }
        }
      },
...
      "OpenCreateTeamLabTopologyModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "networks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologyNetworkModel"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologyAssetModel"
            }
          },
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologyConnectionModel"
            }
          },
          "infrastructure": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologyInfrastructureModel"
            }
          },
          "dependencies": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologyDependencyModel"
            }
          },
          "observation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TeamLabObservationPolicyModel"
              }
            ]
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "OpenTeamLabTopologyPageModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabTopologySummaryModel"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TeamLabValidationResultModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabValidationIssueModel"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "GzctfApiToken": {
        "type": "http",
        "description": "Scoped GZCTF API token",
        "scheme": "bearer",
        "bearerFormat": "GZCTF API token"
      }
    }
  },
...
}